+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [[NEXT]](https://github.com/iExecBlockchainComputing/iexec-common/releases/tag/vNEXT) 2024
 
+### New Features
+
+- Add `createFileNameFromUri` to `FileHashUtils`, use it in `IexecEnvUtils`. (#463)
+
 ### Quality
 
 - Update `HTTPS_FILE_DIGEST` in `FileHelperTests`. (#460)